Migrate all the exercises
