+++ exercises/static/exercises/basic_computer_vision/python_template/ros2_humble/WebGUI.py
@@ -10,7 +10,7 @@ import numpy as np
 # Graphical User Interface Class
 
 
-class GUI(MeasuringThreadingGUI):
+class WebGUI(MeasuringThreadingGUI):
 
     def __init__(self, host="ws://127.0.0.1:2303"):
         super().__init__(host)
