+++ exercises/static/exercises/follow_person/python_template/ros2_humble/WebGUI.py
@@ -10,7 +10,7 @@ from math import cos, sin, atan2
 from gui_interfaces.general.measuring_threading_gui import MeasuringThreadingGUI
 from console_interfaces.general.console import start_console
 
-class GUI(MeasuringThreadingGUI):
+class WebGUI(MeasuringThreadingGUI):
 
     def __init__(self, host="ws://127.0.0.1:2303", freq=30.0):
