+++ exercises/static/exercises/follow_road/python_template/ros2_humble/WebGUI.py
@@ -5,10 +5,10 @@ import threading
 import time
 import numpy as np
 
-from gui_interfaces.general.measuring_threading_gui import MeasuringThreadingGUI
+from gui_interfaces.general.measuring_threading_gui_harmonic import MeasuringThreadingGUI
 from console_interfaces.general.console import start_console
 
-class GUI(MeasuringThreadingGUI):
+class WebGUI(MeasuringThreadingGUI):
 
     def __init__(self, host="ws://127.0.0.1:2303", freq=30.0):
 
@@ -22,6 +22,8 @@ class GUI(MeasuringThreadingGUI):
         self.ack_lock = threading.Lock()
         self.running = True
 
+        self.world_name = "empty"
+
         self.host = host
         self.msg = {"image_right": "", "image_left": ""}
 
@@ -29,6 +31,8 @@ class GUI(MeasuringThreadingGUI):
         self.real_time_factor = 0
         self.frequency_message = {'brain': '', 'gui': '', 'rtf': ''}
         self.iteration_counter = 0
+        self.fps = 0
+        self.lat = 0
 
         self.start()
 
@@ -36,6 +40,7 @@ class GUI(MeasuringThreadingGUI):
     def gui_out_thread(self):
         while self.running:
             start_time = time.time()
+            self.iteration_counter += 1
 
             # Check if a new image should be sent
             with self.ack_lock:
