+++ exercises/static/exercises/marker_visual_loc/python_template/ros2_humble/WebGUI.py
@@ -13,7 +13,7 @@ from console_interfaces.general.console import start_console
 
 from HAL import getPose3d, getOdom
 
-class GUI(MeasuringThreadingGUI):
+class WebGUI(MeasuringThreadingGUI):
 
     def __init__(self, host="ws://127.0.0.1:2303", freq=30.0):
         super().__init__(host)
