+++ exercises/static/exercises/montecarlo_visual_loc/python_template/ros2_humble/WebGUI.py
@@ -12,7 +12,7 @@ from HAL import getPose3d
 
 # Graphical User Interface Class
 
-class GUI(MeasuringThreadingGUI):
+class WebGUI(MeasuringThreadingGUI):
 
     def __init__(self, host="ws://127.0.0.1:2303"):
         super().__init__(host)
