+++ exercises/static/exercises/obstacle_avoidance/python_template/ros2_humble/WebGUI.py
@@ -8,7 +8,7 @@ from HAL import getLaserData, getPose3d
 
 # Graphical User Interface Class
 
-class GUI(MeasuringThreadingGUI):
+class WebGUI(MeasuringThreadingGUI):
 
     def __init__(self, host="ws://127.0.0.1:2303"):
         super().__init__(host)
